♻️ refactor(player): split Interface helpers into hooks/ and contexts/
Mirror the Context/hooks/ pattern inside the Interface domain:

- hooks/      Interface-level layout hooks
                useResolvedGridArea (moved from Interface root)
                useGridTemplate     (moved from src/hooks; only
                                     consumer is Interface/index.tsx
                                     and it pulls AudioPlayer types
                                     and compound metadata directly)
- contexts/   Interface-scoped context data
                playListPortalContext, playListEmptyContext

compound/ stays as a feature group (data + hooks intentionally kept
together). Interface root is now down to index.tsx + Interface.css +
four child component dirs + the three organizational subdirs, and
src/hooks/ is back to truly cross-domain helpers.
