+++ package/src/components/AudioPlayer/Interface/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./useResolvedGridArea";
+export * from "./useGridTemplate";
