+++ package/src/components/AudioPlayer/Interface/hooks/useResolvedGridArea.ts
